fix: Fixed numpy 1.12 indexing and shape preservation

+++ GPy/core/gp_grid.py
@@ -68,12 +68,12 @@ class GpGrid(GP):
         for b in (B.T):
             x = b
             N = 1
-            G = np.zeros(D)
+            G = np.zeros(D, dtype=np.int_)
             for d in range(D):
                 G[d] = len(A[d])
             N = np.prod(G)
             for d in range(D-1, -1, -1):
-                X = np.reshape(x, (G[d], np.round(N/G[d])), order='F')
+                X = np.reshape(x, (G[d], int(np.round(N/G[d]))), order='F')
                 Z = np.dot(A[d], X)
                 Z = Z.T
                 x = np.reshape(Z, (-1, 1), order='F')
